created seperate api routes for student/faculty pages
